Add form checkbox
