+++ frontend/src/metabase/core/components/CheckBox/index.ts
@@ -1,1 +1,2 @@
 export { default } from "./CheckBox";
+export type { CheckBoxProps } from "./CheckBox";
