+++ frontend/src/metabase/core/components/CheckBox/types.ts
@@ -1,5 +1,3 @@
-import { ChangeEvent, FocusEvent, HTMLAttributes, ReactNode } from "react";
-
 export interface CheckBoxInputProps {
   size: number;
 }
@@ -22,27 +20,4 @@ export interface CheckBoxIconContainerProps {
 
 export interface CheckBoxLabelProps {
   labelEllipsis: boolean;
-}
-
-export interface CheckBoxProps
-  extends Omit<HTMLAttributes<HTMLElement>, "onChange" | "onFocus" | "onBlur"> {
-  name?: string;
-  label?: ReactNode;
-  labelEllipsis?: boolean;
-  checked?: boolean;
-  indeterminate?: boolean;
-  disabled?: boolean;
-  size?: number;
-  checkedColor?: string;
-  uncheckedColor?: string;
-  autoFocus?: boolean;
-  onChange?: (event: ChangeEvent<HTMLInputElement>) => void;
-  onFocus?: (event: FocusEvent<HTMLInputElement>) => void;
-  onBlur?: (event: FocusEvent<HTMLInputElement>) => void;
-}
-
-export interface CheckboxTooltipProps {
-  hasTooltip: boolean;
-  tooltipLabel: ReactNode;
-  children: ReactNode;
 }
