+++ frontend/src/metabase/core/components/FormCheckBox/index.ts
@@ -1,0 +1,2 @@
+export { default } from "./FormCheckBox";
+export type { FormCheckBoxProps } from "./FormCheckBox";
