+++ frontend/src/metabase/core/components/FormField/index.ts
@@ -1,1 +1,2 @@
 export { default } from "./FormField";
+export type { FormFieldProps } from "./FormField";
