+++ frontend/src/metabase/core/components/FormField/types.ts
@@ -1,3 +1,2 @@
 export type FormFieldAlignment = "start" | "end";
-
 export type FormFieldOrientation = "horizontal" | "vertical";
