+++ frontend/src/metabase/core/components/FormInput/index.ts
@@ -1,1 +1,2 @@
 export { default } from "./FormInput";
+export type { FormInputProps } from "./FormInput";
